Refactor some stuff and add link to inbox to main menu
