+++ project/src/js/ui/inbox/message/messageView.js
@@ -4,8 +4,6 @@ import layout from '../../layout';
 import * as m from 'mithril';
 import i18n from '../../../i18n';
 import { noop, gameIcon, formatTimeInSecs, formatTournamentDuration, formatTournamentTimeControl } from '../../../utils';
-import faq from '../faq';
-import playerInfo from '../playerInfo';
 import * as helper from '../../helper';
 import settings from '../../../settings';
 import miniBoard from '../../shared/miniBoard';
